# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "post": {
        "operationId": "RoadsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      },
      "get": {
        "operationId": "RoadsController_listProjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      }
    },
    "/projects/{projectId}/roads": {
      "post": {
        "operationId": "RoadsController_createRoad",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      },
      "get": {
        "operationId": "RoadsController_listRoads",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      }
    },
    "/roads/{roadId}/sections": {
      "post": {
        "operationId": "RoadsController_createSection",
        "parameters": [
          {
            "name": "roadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoadSectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      },
      "get": {
        "operationId": "RoadsController_listSections",
        "parameters": [
          {
            "name": "roadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      }
    },
    "/sections/{sectionId}/elements": {
      "post": {
        "operationId": "RoadsController_addElement",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoadElementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      }
    },
    "/sections/{sectionId}/features": {
      "post": {
        "operationId": "RoadsController_addFeature",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoadFeatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      }
    },
    "/sections/{sectionId}/work-items": {
      "post": {
        "operationId": "RoadsController_createWorkItem",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      },
      "get": {
        "operationId": "RoadsController_listWorkItems",
        "parameters": [
          {
            "name": "sectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      }
    },
    "/progress": {
      "post": {
        "operationId": "RoadsController_addProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDailyProgressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roads"
        ]
      }
    },
...
    "/environment/records": {
      "post": {
        "operationId": "EnvironmentController_createRecord",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnvironmentalRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "environment"
        ]
      }
    },
    "/environment/records/{projectId}": {
      "get": {
        "operationId": "EnvironmentController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "environment"
        ]
      }
    },
...
  },
  "info": {
    "title": "RQMS API",
    "description": "Operations & Field Execution API (v1)",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "projectType": {
            "type": "string",
            "enum": [
              "ROAD",
              "BRIDGE",
              "DRAINAGE",
              "OTHER"
            ],
            "default": "ROAD"
          }
        },
        "required": [
          "name",
          "projectType"
        ]
      },
      "CreateRoadDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Unique code within the project."
          },
          "roadType": {
            "type": "string",
            "description": "Free text, e.g. highway, city road."
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
      "CreateRoadSectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startChainage": {
            "type": "number"
          },
          "endChainage": {
            "type": "number"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateRoadElementDto": {
        "type": "object",
        "properties": {
          "elementType": {
            "type": "string",
            "description": "e.g. sidewalk, bike_lane, drainage, lighting"
          },
          "fromChainage": {
            "type": "number"
          },
          "toChainage": {
            "type": "number"
          }
        },
        "required": [
          "elementType"
        ]
      },
      "CreateRoadFeatureDto": {
        "type": "object",
        "properties": {
          "featureType": {
            "type": "string",
            "description": "e.g. inlet, sign, crossing, pole"
          },
          "chainage": {
            "type": "number"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        },
        "required": [
          "featureType"
        ]
      },
      "CreateWorkItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string",
            "description": "Unit of measure, e.g. m, m2, m3, ton, item"
          },
          "roadElementId": {
            "type": "string"
          },
          "roadFeatureId": {
            "type": "string"
          },
          "workLayerId": {
            "type": "string"
          },
          "activityTypeId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "unit"
        ]
      },
      "CreateDailyProgressDto": {
        "type": "object",
        "properties": {
          "workItemId": {
            "type": "string"
          },
          "workDate": {
            "type": "string",
            "description": "YYYY-MM-DD"
          },
          "quantity": {
            "type": "number"
          },
          "fromChainage": {
            "type": "number"
          },
          "toChainage": {
            "type": "number"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          },
          "weather": {
            "type": "object"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "workItemId",
          "workDate",
          "quantity"
        ]
      },
...
      "CreateEnvironmentalRecordDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string",
            "description": "ISO timestamp"
          },
          "parameter": {
            "type": "string",
            "description": "e.g. dust, noise, waste"
          },
          "value": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "projectId",
          "capturedAt",
          "parameter"
        ]
      },
...
    }
  }
}